Add database access to delete command
Logging now can fall back to a table which stores unused channel ids in the
event that no new channels can be made.  This situation will only happen
after deleting a channel from the main server (assuming that the channel is
removed properly), as both servers have the same channel limit.  A message
is sent to the channels both when they are removed and reinstated to help
ease the use for finding old channels' logs.